requirements.txt
appdirs==1.4.4
azure-core==1.29.5
azure-eventhub==5.11.5
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
dotenv==0.0.5
fonttools==4.44.3
frozendict==2.3.8
html5lib==1.1
idna==3.4
install==1.3.5
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.8.1
multitasking==0.0.11
numpy==1.26.2
packaging==23.2
pandas==2.1.3
peewee==3.17.0
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
webencodings==0.5.1
yfinance==0.2.31
